requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
httpie==2.6.0
idna==3.3
psycopg2==2.9.2
Pygments==2.10.0
PyJWT==2.3.0
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
